--- a/proofs/hol_light/arm/Makefile
+++ b/proofs/hol_light/arm/Makefile
@@ -68,7 +68,8 @@ endif
 OBJ = mlkem/mlkem_ntt.o                 \
       mlkem/mlkem_intt.o                \
       mlkem/keccak_f1600_x1_scalar.o    \
-      mlkem/keccak_f1600_x1_v84a.o
+      mlkem/keccak_f1600_x1_v84a.o      \
+      mlkem/keccak_f1600_x2_v84a.o
 
 # According to
 # https://developer.apple.com/documentation/xcode/writing-arm64-code-for-apple-platforms,