--- a/scripts/autogen
+++ b/scripts/autogen
@@ -974,6 +974,9 @@ def gen_hol_light_asm(dry_run=False):
     gen_hol_light_fips202_asm_file(
         "keccak_f1600_x1_v84a_asm.S", "keccak_f1600_x1_v84a.S"
     )
+    gen_hol_light_fips202_asm_file(
+        "keccak_f1600_x2_v84a_asm.S", "keccak_f1600_x2_v84a.S"
+    )
 
 
 def update_via_copy(infile_full, outfile_full, dry_run=False, transform=None):