This commit adds the HOL-Light proof of functional corrrectness
for the v8.4-A+SHA3 based Keccak-f1600-x2 implementation; this
implementation is the default for Arm-based Macs in mlkem-native.

The proof was developed by John Harrison and uses the s2n-bignum
proof infrastructure.

A code change had to be made to align the assembly in mlkem-native
with the one proved correct in HOL-Light:
- Uses of ld2/st2 during loading/storing of the initial/final
  Keccak state are replaced by manual ldp/stp+trn{1,2}.
- An unnecessary stack allocation is removed
- An unnecessary register alias is removed