@hanno-becker 2025-03-04
Only Arm-based Macs are impacted by this change, so here's a local benchmark on my Apple M1:

```
#### Before

$ tests bench -c M1 -r --opt=OPT
INFO  > Benchmark          Compile     (native opt):     make bench OPT=1 AUTO=1 CYCLES=M1 -j10
INFO  > Benchmark          ML-KEM-512  (native opt):     EXEC_WRAPPER=sudo make run_bench_512
INFO  > Benchmark          ML-KEM-512  (native opt):     sudo test/build/mlkem512/bin/bench_mlkem512
   keypair cycles = 12484
    encaps cycles = 15467
    decaps cycles = 20254

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  12457  12468  12474  12477  12480  12484  12488  12492  12496  12504  13568
    encaps percentiles:  15440  15453  15457  15461  15464  15467  15470  15475  15480  15486  16666
    decaps percentiles:  20222  20234  20240  20243  20248  20254  20261  20273  20286  20304  21356

#### After

$ tests bench -c M1 -r --opt=OPT
INFO  > Benchmark          Compile     (native opt):     make bench OPT=1 AUTO=1 CYCLES=M1 -j10
INFO  > Benchmark          ML-KEM-512  (native opt):     EXEC_WRAPPER=sudo make run_bench_512
INFO  > Benchmark          ML-KEM-512  (native opt):     sudo test/build/mlkem512/bin/bench_mlkem512
   keypair cycles = 12370
    encaps cycles = 15396
    decaps cycles = 20133

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  12345  12355  12360  12364  12367  12370  12374  12378  12382  12390  13431
    encaps percentiles:  15369  15379  15385  15388  15392  15396  15398  15402  15408  15413  16580
    decaps percentiles:  20101  20114  20119  20124  20128  20133  20140  20150  20161  20182  21192

#### Before

INFO  > Benchmark          ML-KEM-768  (native opt):     EXEC_WRAPPER=sudo make run_bench_768
INFO  > Benchmark          ML-KEM-768  (native opt):     sudo test/build/mlkem768/bin/bench_mlkem768
   keypair cycles = 21994
    encaps cycles = 24649
    decaps cycles = 31645

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21934  21961  21973  21981  21986  21994  22002  22011  22022  22059  23107
    encaps percentiles:  24594  24617  24630  24637  24643  24649  24657  24665  24680  24710  25758
    decaps percentiles:  31592  31613  31624  31632  31638  31645  31651  31659  31670  31697  32740

#### After

INFO  > Benchmark          ML-KEM-768  (native opt):     EXEC_WRAPPER=sudo make run_bench_768
INFO  > Benchmark          ML-KEM-768  (native opt):     sudo test/build/mlkem768/bin/bench_mlkem768
   keypair cycles = 21799
    encaps cycles = 24451
    decaps cycles = 31444

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  21734  21764  21776  21785  21792  21799  21805  21815  21830  21886  22888
    encaps percentiles:  24388  24416  24429  24439  24445  24451  24458  24467  24483  24516  25533
    decaps percentiles:  31382  31410  31421  31430  31439  31444  31452  31459  31472  31520  32537

#### Before

INFO  > Benchmark          ML-KEM-1024 (native opt):     EXEC_WRAPPER=sudo make run_bench_1024
INFO  > Benchmark          ML-KEM-1024 (native opt):     sudo test/build/mlkem1024/bin/bench_mlkem1024
   keypair cycles = 31176
    encaps cycles = 36366
    decaps cycles = 44953

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  31072  31127  31142  31154  31165  31176  31189  31203  31229  32211  33294
    encaps percentiles:  36237  36290  36312  36328  36348  36366  36385  36406  36441  37469  37620
    decaps percentiles:  44850  44899  44915  44928  44940  44953  44965  44980  45011  45988  46172

#### After

INFO  > Benchmark          ML-KEM-1024 (native opt):     EXEC_WRAPPER=sudo make run_bench_1024
INFO  > Benchmark          ML-KEM-1024 (native opt):     sudo test/build/mlkem1024/bin/bench_mlkem1024
   keypair cycles = 30837
    encaps cycles = 36180
    decaps cycles = 44612

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  30741  30782  30802  30815  30826  30837  30851  30867  30903  31858  32909
    encaps percentiles:  36046  36111  36133  36150  36165  36180  36196  36214  36250  37266  37416
    decaps percentiles:  44511  44555  44575  44587  44600  44612  44630  44648  44679  45631  45779
```